# public/openapi.yaml
openapi: 3.0.3
info:
  title: HookStep / Talent Platform API
  description: REST API for the hiring platform — jobs, applications, profile, and company dashboard.
  version: 1.0.0
servers:
  - url: /
    description: Current origin (relative)
...
paths:
...
  /api/jobs:
    get:
      tags: [Jobs]
      summary: List active jobs (public)
      description: Returns ACTIVE jobs with optional server-side filtering.
      parameters:
        - name: employmentType
          in: query
          description: Comma-separated (e.g. Full-time,Contract,Internship)
          schema: { type: string }
        - name: category
          in: query
          description: e.g. Engineering
          schema: { type: string }
        - name: skills
          in: query
          description: Exact skill tag match (job.skills array contains this value)
          schema: { type: string }
        - name: page
          in: query
          description: Page number (1-based, default 1)
          schema: { type: integer, minimum: 1, default: 1 }
        - name: pageSize
          in: query
          description: Items per page (default 12, max 50)
          schema: { type: integer, minimum: 1, maximum: 50, default: 12 }
      responses:
        '200':
          description: Paginated list of active jobs
          content:
            application/json:
              schema:
                type: object
                required: [jobs, pagination]
                properties:
                  jobs:
                    type: array
                    items: { $ref: '#/components/schemas/JobPublic' }
                  pagination:
                    type: object
                    required: [page, pageSize, total, totalPages, hasMore]
                    properties:
                      page: { type: integer }
                      pageSize: { type: integer }
                      total: { type: integer }
                      totalPages: { type: integer }
                      hasMore: { type: boolean }
    post:
      tags: [Jobs]
      summary: Create job (recruiter only)
      security: [{ session: [] }]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [title, company, description, employmentType, category]
              properties:
                title: { type: string }
                company: { type: string }
                location: { type: string }
                description: { type: string }
                employmentType: { type: string }
                category: { type: string }
                skills: { type: array, items: { type: string } }
                salary: { type: string }
                currency: { type: string }
                status: { type: string, enum: [ACTIVE, DRAFT] }
                experienceMin: { type: integer }
                experienceMax: { type: integer }
                country: { type: string }
                state: { type: string }
                city: { type: string }
      responses:
        '201':
          description: Job created
          content:
            application/json:
              schema: { $ref: '#/components/schemas/Job' }
        '401': { $ref: '#/components/responses/Unauthorized' }
        '403': { $ref: '#/components/responses/Forbidden' }
        '400': { $ref: '#/components/responses/BadRequest' }
...
components:
  securitySchemes:
    session:
      type: apiKey
      in: cookie
      name: next-auth.session-token
      description: NextAuth session cookie (sign in via the app)
  responses:
    Unauthorized:
      description: Not authenticated
      content:
        application/json:
          schema: { $ref: '#/components/schemas/Error' }
      headers: {}
    Forbidden:
      description: Authenticated but not allowed (e.g. wrong role)
      content:
        application/json:
          schema: { $ref: '#/components/schemas/Error' }
    BadRequest:
      description: Invalid request
      content:
        application/json:
          schema: { $ref: '#/components/schemas/Error' }
...
  schemas:
    Error:
      type: object
      properties:
        error: { type: string }
    Job:
      type: object
      properties:
        id: { type: string }
        title: { type: string }
        company: { type: string }
        location: { type: string }
        description: { type: string }
        employmentType: { type: string }
        category: { type: string }
        skills: { type: array, items: { type: string } }
        salary: { type: string }
        currency: { type: string }
        status: { type: string }
        experienceMin: { type: integer }
        experienceMax: { type: integer }
        createdAt: { type: string, format: date-time }
        updatedAt: { type: string, format: date-time }
    JobPublic:
      type: object
      description: Same as Job but only safe/public fields (no companyId)
      properties:
        id: { type: string }
        title: { type: string }
        company: { type: string }
        location: { type: string }
        description: { type: string }
        employmentType: { type: string }
        category: { type: string }
        skills: { type: array, items: { type: string } }
        salary: { type: string }
        currency: { type: string }
        status: { type: string }
        experienceMin: { type: integer }
        experienceMax: { type: integer }
        createdAt: { type: string, format: date-time }
...